schedule_app/backend/requirements.txt
# Django and Django REST Framework
Django==5.1.7
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.1
networkx==3.2.1
numpy==1.26.4
Pillow==10.2.0

# Development
black==24.3.0
flake8==7.0.0